Update profiling scripts
