+++ profiling/custom_shapes.py
@@ -25,7 +25,7 @@ def draw():
             transformed_pts = transform(points, np.asarray((20, 20)), np.asarray((width / 2, height / 2)))
             for p in transformed_pts:
                 vertex(*p)
-            end_shape("CLOSED")
+            end_shape("CLOSE")
 
 
 if __name__ == '__main__':
